task-management-service/pyproject.toml
[project]
name = "task-management-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.27",
    "asyncpg>=0.30.0",
    "alembic>=1.17.1",
    "greenlet>=3.2.4",
    "psycopg[binary]>=3.2.12",
    "psycopg2-binary>=2.9.11",
    "google-cloud-pubsub>=2.33.0",
    "pydantic-settings>=2.11.0",
]
